% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ca0_fun.R
\name{Ca0_fun}
\alias{Ca0_fun}
\title{Calculating normalising constant for an NPP}
\usage{
Ca0_fun(X0 = X0, Y0 = Y0, Z0 = Z0, a0_increment = a0_increment)
}
\arguments{
\item{X0}{A matrix. The design matrix for the historical dataset, excluding the intercept term. The first column must represent treatment allocation. Passed from \link[PPCRCT]{NPP}.}

\item{Y0}{A vector containing the outcome data for the current dataset. Passed from \link[PPCRCT]{NPP}.}

\item{Z0}{A vector of consecutive integers containing cluster indices for the historical dataset. Passed from \link[PPCRCT]{NPP}.}

\item{sigmaprior}{One of either "hnormal" or "hcauchy" to indicated whethere a half-normal or half-cauchy prior distribution should be fitted to the between-cluster SD parameter. Passed from \link[PPCRCT]{NPP}.}

\item{reg.prior.mean}{The mean for the normal prior distribution for each of the regression coefficients. Passed from \link[PPCRCT]{NPP}.}

\item{reg.prior.sd}{The standard deviation for the normal prior distribution for each of the regression coefficients. Passed from \link[PPCRCT]{NPP}.}

\item{nits_normalise}{An integer. Number of iterations per chain used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{NPP}.}

\item{burnin_normalise}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{NPP}.}

\item{nchains_normalise}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{NPP}.}

\item{max_treedepth_normalise}{Maximum treedepth for the Markov Chain Monte Carlo procedure for estimating the normalising constant. See link stan documentation. Passed from \link[PPCRCT]{NPP}.}

\item{thin_normalise}{A positive integer specifying the period for saving Markov Chain Monte Carlo samples for the procedure estimating the normalising constant. Defaults to 1. Passed from \link[PPCRCT]{NPP}.}

\item{adapt_delta_normalise}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. See \link[rstan]{sampling} documentation. Passed from \link[PPCRCT]{NPP}.}

\item{adapt_delta}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the NPP. See link stan documentation}
}
\description{
Ca0_fun is a modularised function contained within \link[PPCRCT]{NPP} function. Not for use by users.
}
