% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDPP.R
\name{FDPP}
\alias{FDPP}
\title{Fitting Fixed Discounting Power Priors}
\usage{
FDPP()
}
\arguments{
\item{X}{A matrix. The design matrix for the current dataset, excluding the intercept term. The first column must represent treatment allocation.}

\item{X0}{A matrix. The design matrix for the historical dataset, excluding the intercept term. The first column must represent treatment allocation.}

\item{Y}{A vector containing the outcome data for the current dataset}

\item{Y0}{A vector containing the outcome data for the current dataset}

\item{Z}{A vector of consecutive integers containing cluster indices for the current dataset.}

\item{Z0}{A vector of consecutive integers containing cluster indices for the historical dataset.}

\item{sigma.b.prior}{One of either "hnormal" or "hcauchy" to indicated whethere a half-normal or half-cauchy prior distribution should be fitted to the between-cluster SD parameter}

\item{intercept.prior.mean}{The mean for the normal prior distribution for the intercept}

\item{intercept.prior.sd}{The standard deviation for the normal prior distribution for the intercept}

\item{reg.prior.mean}{A vector of means for the normal prior distribution for each of the regression coefficients (of length equal to the number of columns of \code{X0}).}

\item{reg.prior.sd}{A vector of standard deviations for the normal prior distribution for each of the regression coefficients(of length equal to the number of columns of \code{X0}).}

\item{sigma.b.prior.parm}{The parameter for the prior distribution on the between cluster standard deviation. If \code{sigma.b.prior = "hcauchy"} this represents the scale parameter of the half-cauchy distribution. If \code{sigma.b.prior = "hnormal"} this represents that standard deviation parameter of the half-normal distribution.}

\item{sigma.prior.parm}{The rate parameter for the exponential prior distribution for the residual standard deviation.}

\item{nits_fdpp}{An integer. Number of iterations per chain to be used in the Markov Chain Monte Carlo procedure for fitting the NPP model}

\item{burnin_fdpp}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for fitting the NPP model}

\item{nchains_fdpp}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant}

\item{max_treedepth_fdpp}{Maximum treedepth for the Markov Chain monte carlo procedure for estimating the NPP. See link stan documentation}

\item{thin_fdpp}{A positive integer specifying the period for saving Markov Chain Monte Carlo samples for the procedure fitting the NPP model. Defaults to 1}

\item{adapt_delta_npp}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the NPP. See link stan documentation}

\item{a0_increment}{Value of the increments by which \code{a0} is increased between each estimation of the normalising constant.}

\item{seed}{Set the seed.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
TO UPDATE
}
\description{
FDPP is used to fit a Fixed Discounting Power Prior (FDPP) to analysis of a (current) dataset, using a second (historical) dataset to formulate the power prior.
}
\examples{
TO UPDATE;
}
