% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDPP_modelfit.R
\name{FDPP_modelfit}
\alias{FDPP_modelfit}
\title{Fit the Fixed Discounting Power Prior in Stan}
\usage{
FDPP_modelfit(
  X = X,
  X0 = X0,
  Y = Y,
  Y0 = Y0,
  Z = Z,
  Z0 = Z0,
  a0 = a0,
  sigma.b.prior = sigma.b.prior,
  intercept.prior.mean = intercept.prior.mean,
  intercept.prior.sd = intercept.prior.sd,
  reg.prior.mean = reg.prior.mean,
  reg.prior.sd = reg.prior.sd,
  sigma.b.prior.parm = sigma.b.prior.parm,
  sigma.prior.parm = sigma.prior.parm,
  nits_fdpp = 5000,
  burnin_fdpp = NULL,
  nchains_fdpp = 4,
  max_treedepth_fdpp = 10,
  thin_fdpp = 1,
  adapt_delta_fdpp = 0.95,
  seed = seed,
  cores = cores
)
}
\arguments{
\item{X}{A matrix. The design matrix for the current dataset, excluding the intercept term. The first column must represent treatment allocation.}

\item{X0}{A matrix. The design matrix for the historical dataset, excluding the intercept term. The first column must represent treatment allocation. Passed from \link[PPCRCT]{FDPP}.}

\item{Y}{A vector containing the outcome data for the current dataset}

\item{Y0}{A vector containing the outcome data for the current dataset. Passed from \link[PPCRCT]{FDPP}.}

\item{Z}{A vector of consecutive integers containing cluster indices for the current dataset.}

\item{Z0}{A vector of consecutive integers containing cluster indices for the historical dataset. Passed from \link[PPCRCT]{FDPP}.}

\item{a0}{The discounting factor. Must be a value between 0 and 1.}

\item{sigma.b.prior}{One of either "hnormal" or "hcauchy" to indicated whethere a half-normal or half-cauchy prior distribution should be fitted to the between-cluster SD parameter}

\item{intercept.prior.mean}{The mean for the normal prior distribution for the intercept}

\item{intercept.prior.sd}{The standard deviation for the normal prior distribution for the intercept}

\item{reg.prior.mean}{A vector of means for the normal prior distribution for each of the regression coefficients (of length equal to the number of columns of \code{X0}). Passed from \link[PPCRCT]{FDPP}.}

\item{reg.prior.sd}{A vector of standard deviations for the normal prior distribution for each of the regression coefficients(of length equal to the number of columns of \code{X0}). Passed from \link[PPCRCT]{FDPP}.}

\item{sigma.b.prior.parm}{The parameter for the prior distribution on the between cluster standard deviation. If \code{sigma.b.prior = "hcauchy"} this represents the scale parameter of the half-cauchy distribution. If \code{sigma.b.prior = "hnormal"} this represents that standard deviation parameter of the half-normal distribution.}

\item{sigma.prior.parm}{The rate parameter for the exponential prior distribution for the residual standard deviation.  Passed from \link[PPCRCT]{FDPP}.}

\item{seed}{Set the seed.}

\item{cores}{Number of cores to use in MCMC procedure.}

\item{nits_normalise}{An integer. Number of iterations per chain used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{FDPP}.}

\item{burnin_normalise}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{FDPP}.}

\item{nchains_normalise}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Passed from \link[PPCRCT]{FDPP}.}

\item{max_treedepth_normalise}{Maximum treedepth for the Markov Chain Monte Carlo procedure for estimating the normalising constant. See link stan documentation. Passed from \link[PPCRCT]{FDPP}.}

\item{thin_normalise}{A positive integer specifying the period for saving Markov Chain Monte Carlo samples for the procedure estimating the normalising constant. Defaults to 1. Passed from \link[PPCRCT]{FDPP}.}

\item{adapt_delta_normalise}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. See \link[rstan]{sampling}. Passed from \link[PPCRCT]{FDPP}.}
}
\value{
Returns a grid of values of \code{a0} between 0 and 1 of length 10000, and associated estimates of the normalising constant.
}
\description{
FDPP_modelfit is a modularised function contained within \link[PPCRCT]{FDPP} function, which fits the Fixed Discount Power Prior (i.e. information borrowing from all parameters) in Stan.
}
