% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{NPP}
\alias{NPP}
\title{Fitting Normalised Power Priors}
\usage{
NPP(
  X,
  X0,
  Y,
  Y0,
  Z,
  Z0,
  sigma.b.prior = c("hnormal", "hcauchy"),
  intercept.prior.mean = 0,
  intercept.prior.sd = NULL,
  reg.prior.mean = 0,
  reg.prior.sd = NULL,
  sigma.b.prior.parm = NULL,
  sigma.prior.parm = NULL,
  nits_normalise = 2000,
  burnin_normalise = NULL,
  nchains_normalise = 4,
  max_treedepth_normalise = 10,
  thin_normalise = 1,
  adapt_delta_normalise = 0.95,
  nits_npp = 5000,
  burnin_npp = NULL,
  nchains_npp = 4,
  max_treedepth_npp = 10,
  thin_npp = 1,
  adapt_delta_npp = 0.95,
  a0_increment = 0.05,
  seed = 12345,
  parallel = F,
  ...
)
}
\arguments{
\item{X}{A matrix. The design matrix for the current dataset, excluding the intercept term. The first column must represent treatment allocation.}

\item{X0}{A matrix. The design matrix for the historical dataset, excluding the intercept term. The first column must represent treatment allocation.}

\item{Y}{A vector containing the outcome data for the current dataset.}

\item{Y0}{A vector containing the outcome data for the historical dataset.}

\item{Z}{A vector of consecutive integers containing cluster indices for the current dataset.}

\item{Z0}{A vector of consecutive integers containing cluster indices for the historical dataset.}

\item{sigma.b.prior}{One of either "hnormal" or "hcauchy" to indicate whether a half-normal or half-cauchy prior distribution should be fitted to the between-cluster SD parameter}

\item{intercept.prior.mean}{The mean for the normal prior distribution for the intercept. Defaults to 0.}

\item{intercept.prior.sd}{The standard deviation for the normal prior distribution for the intercept}

\item{reg.prior.mean}{A vector of means for the normal prior distribution for each of the regression coefficients (of length equal to the number of columns of \code{X0}).}

\item{reg.prior.sd}{A vector of standard deviations for the normal prior distribution for each of the regression coefficients(of length equal to the number of columns of \code{X0}).}

\item{sigma.b.prior.parm}{The parameter for the prior distribution on the between-cluster standard deviation. If \code{sigma.b.prior = "hcauchy"} this represents the scale parameter of the Half-Cauchy distribution. If \code{sigma.b.prior = "hnormal"} this represents that standard deviation parameter of the Half-Normal distribution.}

\item{sigma.prior.parm}{The rate parameter for the exponential prior distribution for the residual standard deviation.}

\item{nits_normalise}{An integer. Number of iterations per chain used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Defaults to 2000. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{burnin_normalise}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for estimating the normalising constant. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{nchains_normalise}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Defaults to 4. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{max_treedepth_normalise}{An integer. Maximum treedepth for the Markov Chain Monte Carlo procedure for estimating the normalising constant. Defaults to 10. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{thin_normalise}{A positive integer specifying the period for saving Markov Chain Monte Carlo samples for the procedure estimating the normalising constant. Defaults to 1. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{adapt_delta_normalise}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. Defaults to 0.95. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{nits_npp}{An integer. Number of iterations per chain to be used in the Markov Chain Monte Carlo procedure for fitting the NPP model. Defaults to 5000. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{burnin_npp}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for fitting the NPP model. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{nchains_npp}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for fitting the NPP model. Defaults to 4. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{max_treedepth_npp}{Maximum treedepth for the Markov Chain monte carlo procedure for fitting the NPP model. Defaults to 10. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{thin_npp}{A positive integer specifying the period for saving Markov Chain Monte Carlo samples for the procedure fitting the NPP model. Defaults to 1. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{adapt_delta_npp}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for fitting the NPP model. Defaults to 0.95. See \code{\link[rstan:stan]{rstan::stan()}} for further details.}

\item{a0_increment}{Value of the increments by which \code{a0} is increased between each estimation of the normalising constant. Defaults to 0.05.}

\item{seed}{Set the seed.}

\item{parallel}{logical. If TRUE, parallelisation of MCMC chains is implemented.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of S4 class stanfit representing the fitted results. \code{beta[1]} represents the treatment effect parameter.
}
\description{
NPP is used to fit a Normalised Power Prior (NPP) to analysis of a (current) dataset, using a second (historical) dataset to formulate the power prior, where both datasets contain clustering.
}
