% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{NPP}
\alias{NPP}
\title{Fitting Normalised Power Priors}
\usage{
NPP(
  Trt,
  Trt0,
  X,
  X0,
  Y,
  Y0,
  sigmaprior = c("hnormal", "hcauchy"),
  reg.prior.mean = 0,
  reg.prior.sd,
  nits = 2000,
  burnin = nits/2,
  max_treedepth = 10,
  adapt_delta = 0.8
)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
NPP is used to fit a Normalised Power Prior (NPP) to analysis of a dataset, using a second (historical) dataset to formulate the power prior.
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
