% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{NPP}
\alias{NPP}
\title{Fitting Normalised Power Priors}
\usage{
NPP(
  Trt,
  Trt0,
  X = NULL,
  X0 = NULL,
  Y,
  Y0,
  Z,
  Z0,
  sigmaprior = c("hnormal", "hcauchy"),
  reg.prior.mean = 0,
  reg.prior.sd,
  nits_normalise = 2000,
  burnin_normalise = nits_normalise/2,
  nchains_normalise = 4,
  max_treedepth_normalise = 10,
  adapt_delta_normalise = 0.95,
  nits = 5000,
  burnin = nits/2,
  nchains = 4,
  max_treedepth = 10,
  adapt_delta = 0.95,
  ...
)
}
\arguments{
\item{Trt}{A vector of 1's and 0's denoting treatment or control arms, respectively, for the current dataset}

\item{Trt0}{A vector of 1's and 0's denoting treatment or control arms, respectivly, for the historical dataset}

\item{X}{A matrix. The design matrix for the current dataset, excluding treatment allocation indicator and intercept term}

\item{X0}{A matrix. The design matrix for the historical dataset, excluding treatment allocation indicator and intercept term}

\item{Y}{A vector containing the outcome data for the current dataset}

\item{Y0}{A vector containing the outcome data for the current dataset}

\item{Z}{A vector of consecutive integers containing cluster indices for the current dataset.}

\item{Z0}{A vector of consecutive integers containing cluster indices for the historical dataset.}

\item{sigmaprior}{One of either "hnormal" or "hcauchy" to indicated whethere a half-normal or half-cauchy prior distribution should be fitted to the between-cluster SD parameter}

\item{reg.prior.mean}{The mean for the normal prior distribution for each of the regression coefficients}

\item{reg.prior.sd}{The standard deviation for the normal prior distribution for each of the regression coefficients}

\item{nits_normalise}{An integer. Number of iterations per chain used in the Markov Chain Monte Carlo procedure for estimating the normalising constant}

\item{burnin_normalise}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for estimating the normalising constant}

\item{nchains_normalise}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant}

\item{max_treedepth_normalise}{Maximum treedepth for the Markov Chain monte carlo procedure for estimating the normalising constant. See link stan documentation}

\item{adapt_delta_normalise}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the normalising constant. See link stan documentation}

\item{nits}{An integer. Number of iterations per chain to be used in the Markov Chain Monte Carlo procedure for fitting the NPP model}

\item{burnin}{An integer. Number of iterations per chain to be discarded in the Markov Chain Monte Carlo procedure for fitting the NPP model}

\item{nchains}{An integer. Number of chains to be used in the Markov Chain Monte Carlo procedure for estimating the normalising constant}

\item{max_treedepth}{Maximum treedepth for the Markov Chain monte carlo procedure for estimating the NPP. See link stan documentation}

\item{adapt_delta}{Value of adapt delta used in the Markov Chain Monte Carlo procedure for estimating the NPP. See link stan documentation}

\item{...}{Further arguments passed to or from other methods}
}
\value{
TO UPDATE
}
\description{
NPP is used to fit a Normalised Power Prior (NPP) to analysis of a (crruent) dataset, using a second (historical) dataset to formulate the power prior.
}
\examples{
TO UPDATE;
}
